+++ packages/create-pantheon-decoupled-kit/__tests__/addDependencies.test.ts
@@ -1,26 +1,19 @@
-import * as actions from '../src/actions/index';
 import fs from 'fs-extra';
 import path from 'path';
 import process from 'process';
+import * as actions from '../src/actions/index';
+import { sharedTestData } from './sharedTestData';
 
 vi.mock('inquirer');
 
-const pkgPath = path.resolve(
+const outDir = path.resolve(
 	process.cwd(),
 	'__tests__',
 	'templates',
 	'addDependencies',
-	'package.json',
 );
 
-const globalData = {
-	_: ['test-diff'],
-	// 'workaround' for tests to work
-	templateRootDir: `${process.cwd()}/__tests__/`,
-	outDir: `${process.cwd()}/__tests__/templates/addDependencies`,
-	force: false,
-	silent: false,
-};
+const pkgPath = path.resolve(outDir, 'package.json');
 
 const emptyPkgJson = {
 	name: 'add-dependencies',
@@ -31,7 +24,7 @@ const emptyPkgJson = {
 	license: 'ISC',
 };
 
-describe('addWithDiff()', () => {
+describe('addDependencies()', () => {
 	beforeEach((context) => {
 		fs.ensureFileSync(pkgPath);
 		fs.writeFileSync(pkgPath, JSON.stringify(emptyPkgJson));
@@ -49,8 +42,11 @@ describe('addWithDiff()', () => {
 		addDepsSpy,
 		readFileSpy,
 	}) => {
-		const data = Object.assign({}, globalData);
+		const data = Object.assign({}, sharedTestData);
+		data.outDir = outDir;
+
 		actions.addDependencies({ data });
+
 		expect(readFileSpy).toHaveBeenCalled();
 		expect(readFileSpy).toHaveReturnedWith(JSON.stringify(emptyPkgJson));
 		expect(addDepsSpy).toHaveBeenCalled();
@@ -59,12 +55,13 @@ describe('addWithDiff()', () => {
 	it('should write to the package.json of the outDir with dependencies in `data`', ({
 		addDepsSpy,
 	}) => {
-		const data = Object.assign(
-			{ dependencies: { 'test-dep': '0.0.0' } },
-			globalData,
-		);
+		const data = Object.assign({}, sharedTestData);
+		data.dependencies = { 'test-dep': '0.0.0' };
+		data.outDir = outDir;
+
 		actions.addDependencies({ data });
 		const result = JSON.parse(fs.readFileSync(pkgPath, 'utf-8'));
+
 		expect(result.dependencies).toEqual(data.dependencies);
 		expect(addDepsSpy).toHaveBeenCalled();
 	});
@@ -72,29 +69,34 @@ describe('addWithDiff()', () => {
 	it('should write to the package.json of the outDir with devDependencies in `data`', ({
 		addDepsSpy,
 	}) => {
-		const data = Object.assign(
-			{ devDependencies: { 'test-dep': '0.0.0' } },
-			globalData,
-		);
+		const data = Object.assign({}, sharedTestData);
+		data.devDependencies = { 'test-dep': '0.0.0' };
+		data.outDir = outDir;
+
 		actions.addDependencies({ data });
+
 		const result = JSON.parse(fs.readFileSync(pkgPath, 'utf-8'));
 		expect(result.devDependencies).toEqual(data.devDependencies);
 		expect(addDepsSpy).toHaveBeenCalled();
 	});
 
 	it.fails('should throw an error if outDir is not valid', ({ addDepsSpy }) => {
-		const data = Object.assign({}, globalData);
+		const data = Object.assign({}, sharedTestData);
 		data.outDir = '';
+
 		actions.addDependencies({ data });
+
 		expect(addDepsSpy).toThrowError('outDir is not valid');
 	});
 
 	it.fails(
 		'should throw an error if package.json is not found in outDir',
 		({ addDepsSpy }) => {
-			const data = Object.assign({}, globalData);
+			const data = Object.assign({}, sharedTestData);
 			data.outDir = `${process.cwd()}/__tests__`;
+
 			actions.addDependencies({ data });
+
 			expect(addDepsSpy).toThrowError();
 		},
 	);