+++ packages/create-pantheon-decoupled-kit/__tests__/runLint.test.ts
@@ -1,8 +1,8 @@
-import * as actions from '../src/actions/runLint';
 import chalk from 'chalk';
-import whichPMRuns from 'which-pm-runs';
 import { execSync } from 'child_process';
-import type { ParsedArgs } from 'minimist';
+import whichPMRuns from 'which-pm-runs';
+import * as actions from '../src/actions/runLint';
+import { globalData } from './sharedTestData';
 
 vi.mock('child_process');
 vi.mock('which-pm-runs');
@@ -10,18 +10,30 @@ vi.mock('which-pm-runs');
 const outDir = (dir: 'withLint' | 'withoutLint' | 'withLintFix') =>
 	`${process.cwd()}/__tests__/fixtures/runESLint/${dir}`;
 
+const dataWithLint = {
+	...globalData,
+	outDir: outDir('withLint'),
+};
+const dataWithLintFix = {
+	...globalData,
+	outDir: outDir('withLintFix'),
+};
+const dataWithoutLint = {
+	...globalData,
+	outDir: outDir('withoutLint'),
+};
+
 describe('runEsLint()', () => {
 	afterEach(() => {
 		vi.resetAllMocks();
 	});
 	it('should lint and format the outDir using the detected package manager: pnpm', async () => {
-		const data: ParsedArgs = { _: [], outDir: outDir('withLintFix') };
 		vi.mocked(whichPMRuns).mockImplementationOnce(() => ({
 			name: 'pnpm',
 			version: 'x',
 		}));
 
-		await actions.runLint({ data });
+		await actions.runLint({ data: dataWithLintFix });
 		expect(vi.mocked(execSync)).toHaveBeenCalledWith('pnpm lint:fix', {
 			stdio: 'inherit',
 			cwd: outDir('withLintFix'),
@@ -30,14 +42,12 @@ describe('runEsLint()', () => {
 	});
 
 	it('should lint and format the outDir using the detected package manager: npm', async () => {
-		const data: ParsedArgs = { _: [], outDir: outDir('withLintFix') };
-
 		vi.mocked(whichPMRuns).mockImplementationOnce(() => ({
 			name: 'npm',
 			version: 'x',
 		}));
 
-		await actions.runLint({ data });
+		await actions.runLint({ data: dataWithLintFix });
 		expect(vi.mocked(execSync)).toHaveBeenCalledWith('npm run lint:fix', {
 			stdio: 'inherit',
 			cwd: outDir('withLintFix'),
@@ -46,14 +56,12 @@ describe('runEsLint()', () => {
 	});
 
 	it('should lint and format the outDir with no lint script installed', async () => {
-		const data: ParsedArgs = { _: [], outDir: outDir('withoutLint') };
-
 		vi.mocked(whichPMRuns).mockImplementationOnce(() => ({
 			name: 'npm',
 			version: 'x',
 		}));
 
-		await actions.runLint({ data });
+		await actions.runLint({ data: dataWithoutLint });
 		expect(vi.mocked(execSync)).toHaveBeenCalledWith(
 			'npx eslint --ext {js,ts} --fix',
 			{
@@ -65,14 +73,12 @@ describe('runEsLint()', () => {
 	});
 
 	it('should use the lint command if it is found but not lint:fix script is available', async () => {
-		const data: ParsedArgs = { _: [], outDir: outDir('withLint') };
-
 		vi.mocked(whichPMRuns).mockImplementationOnce(() => ({
 			name: 'npm',
 			version: 'x',
 		}));
 
-		await actions.runLint({ data });
+		await actions.runLint({ data: dataWithLint });
 		expect(vi.mocked(execSync)).toHaveBeenCalledWith('npm run lint', {
 			stdio: 'inherit',
 			cwd: outDir('withLint'),
@@ -81,7 +87,7 @@ describe('runEsLint()', () => {
 	});
 
 	it.fails('should throw if outDir is not defined', async () => {
-		const data: ParsedArgs = { _: [] };
+		const data = Object.assign({ _: [] }, globalData);
 
 		const runESLintSpy = vi.spyOn(actions, 'runLint');
 
@@ -90,15 +96,13 @@ describe('runEsLint()', () => {
 	});
 
 	it.fails('should throw if there is an error', async () => {
-		const data: ParsedArgs = { _: [], outDir: outDir('withLint') };
-
 		const runESLintSpy = vi.spyOn(actions, 'runLint');
 		const errorLogSpy = vi.spyOn(console, 'error');
 		vi.mocked(execSync).mockImplementationOnce(() => {
 			throw new Error('Some error happened');
 		});
 
-		await actions.runLint({ data });
+		await actions.runLint({ data: dataWithLint });
 		expect(errorLogSpy).toHaveBeenCalledWith(
 			chalk.red('There was a problem linting:'),
 		);