+++ packages/create-pantheon-decoupled-kit/CHANGELOG.md
@@ -1,5 +1,19 @@
 # create-pantheon-decoupled-kit
 
+## 0.10.4-canary.1
+
+### Patch Changes
+
+- da7ebb26: Bump `@pantheon-systems/wordpress-kit` and
+  `@pantheon-systems/drupal-kit`
+
+## 0.10.4-canary.0
+
+### Patch Changes
+
+- 5cf51bed: Some internal changes to types for better API reference generation
+- e791c027: Bump `@pantheon-systems/decoupled-kit-health-check` version
+
 ## 0.10.3
 
 ### Patch Changes